Refactor ImageStructGenerator
