+++ Sources/RswiftCore/Generators/ImageStructGenerator.swift
@@ -58,24 +58,27 @@ struct ImageStructGenerator: ExternalOnlyStructGenerator {
     }
 
     return Struct(
-      comments: ["This `R.image` struct is generated, and contains static references to \(imageLets.count) images."],
+      comments: ["This `\(qualifiedName)` struct is generated, and contains static references to \(imageLets.count) images."],
       accessModifier: externalAccessLevel,
-      type: Type(module: .host, name: "image"),
+      type: Type(module: .host, name: structName),
       implements: [],
       typealiasses: [],
       properties: imageLets,
-      functions: groupedFunctions.uniques.map { imageFunction(for: $0, at: externalAccessLevel) },
+      functions: groupedFunctions.uniques.map { imageFunction(for: $0, at: externalAccessLevel, prefix: qualifiedName) },
       structs: structs,
       classes: []
     )
   }
 
-  private func imageFunction(for name: String, at externalAccessLevel: AccessLevel) -> Function {
+  private func imageFunction(for name: String, at externalAccessLevel: AccessLevel, prefix: SwiftIdentifier) -> Function {
+    let structName = SwiftIdentifier(name: name)
+    let qualifiedName = prefix + structName
+
     return Function(
       comments: ["`UIImage(named: \"\(name)\", bundle: ..., traitCollection: ...)`"],
       accessModifier: externalAccessLevel,
       isStatic: true,
-      name: SwiftIdentifier(name: name),
+      name: structName,
       generics: nil,
       parameters: [
         Function.Parameter(
@@ -87,7 +90,7 @@ struct ImageStructGenerator: ExternalOnlyStructGenerator {
       ],
       doesThrow: false,
       returnType: Type._UIImage.asOptional(),
-      body: "return UIKit.UIImage(resource: R.image.\(SwiftIdentifier(name: name)), compatibleWith: traitCollection)"
+      body: "return UIKit.UIImage(resource: \(qualifiedName), compatibleWith: traitCollection)"
     )
   }
 }
@@ -132,18 +135,21 @@ extension NamespacedAssetSubfolder: ExternalOnlyStructGenerator {
       implements: [],
       typealiasses: [],
       properties: imageLets,
-      functions: groupedFunctions.uniques.map { imageFunction(for: $0, at: externalAccessLevel) },
+      functions: groupedFunctions.uniques.map { imageFunction(for: $0, at: externalAccessLevel, prefix: qualifiedName) },
       structs: structs,
       classes: []
     )
   }
 
-  private func imageFunction(for name: String, at externalAccessLevel: AccessLevel) -> Function {
+  private func imageFunction(for name: String, at externalAccessLevel: AccessLevel, prefix: SwiftIdentifier) -> Function {
+    let structName = SwiftIdentifier(name: name)
+    let qualifiedName = prefix + structName
+
     return Function(
       comments: ["`UIImage(named: \"\(name)\", bundle: ..., traitCollection: ...)`"],
       accessModifier: externalAccessLevel,
       isStatic: true,
-      name: SwiftIdentifier(name: name),
+      name: structName,
       generics: nil,
       parameters: [
         Function.Parameter(
@@ -155,35 +161,7 @@ extension NamespacedAssetSubfolder: ExternalOnlyStructGenerator {
       ],
       doesThrow: false,
       returnType: Type._UIImage.asOptional(),
-      body: "return UIKit.UIImage(resource: R.image.\(path).\(SwiftIdentifier(name: name)), compatibleWith: traitCollection)"
+      body: "return UIKit.UIImage(resource: \(qualifiedName), compatibleWith: traitCollection)"
     )
-  }
-}
-
-struct AssetSubfolders {
-  let folders: [NamespacedAssetSubfolder]
-  let duplicates: [NamespacedAssetSubfolder]
-
-  init(all subfolders: [NamespacedAssetSubfolder], assetIdentifiers: [SwiftIdentifier]) {
-    var dict: [SwiftIdentifier: NamespacedAssetSubfolder] = [:]
-
-    for subfolder in subfolders {
-      let name = SwiftIdentifier(name: subfolder.name)
-      if let duplicate = dict[name] {
-        duplicate.subfolders += subfolder.subfolders
-        duplicate.imageAssets += subfolder.imageAssets
-      } else {
-        dict[name] = subfolder
-      }
-    }
-
-    self.folders = dict.values.filter { !assetIdentifiers.contains(SwiftIdentifier(name: $0.name)) }
-    self.duplicates = dict.values.filter { assetIdentifiers.contains(SwiftIdentifier(name: $0.name)) }
-  }
-
-  func printWarningsForDuplicates() {
-    for subfolder in duplicates {
-      warn("Skipping asset subfolder because symbol '\(subfolder.name)' would conflict with image: \(subfolder.name)")
-    }
   }
 }
